+++ grafana-plugin/src/containers/UserSettings/parts/tabs/UserInfoTab/UserInfoTab.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
 
-import { Label } from '@grafana/ui';
-import cn from 'classnames/bind';
+import { InlineField, Input, Legend } from '@grafana/ui';
 
-import Text from 'components/Text/Text';
 import { UserSettingsTab } from 'containers/UserSettings/UserSettings.types';
 import { Connectors } from 'containers/UserSettings/parts/connectors';
 import { User } from 'models/user/user.types';
 import { useStore } from 'state/useStore';
-
-import styles from './UserInfoTab.module.css';
-
-const cx = cn.bind(styles);
 
 interface UserInfoTabProps {
   id: User['pk'];
@@ -25,23 +19,27 @@ export const UserInfoTab = (props: UserInfoTabProps) => {
   const { userStore } = store;
 
   const storeUser = userStore.items[id];
+  let width = 12;
 
   return (
     <>
-      <div className={cx('user-item')}>
-        <Text type="secondary">
-          To edit user details such as Username, email, and roles, please visit{' '}
-          <a href="/org/users"> Grafana User settings</a>.
-        </Text>
-      </div>
-      <div className={cx('user-item')}>
-        <Label>Username:</Label>
-        <span className={cx('user-value')}>{storeUser.username || '—'}</span>
-      </div>
-      <div className={cx('user-item')}>
-        <Label>Email:</Label>
-        <span className={cx('user-value')}>{storeUser.email || '—'}</span>
-      </div>
+      <Legend>User information</Legend>
+      <InlineField
+        label="Username"
+        labelWidth={width}
+        grow
+        disabled
+        tooltip="To edit username go to Grafana user management"
+      >
+        <Input value={storeUser.username || ''} />
+      </InlineField>
+      <InlineField label="Email" labelWidth={width} grow disabled tooltip="To edit email go to Grafana user management">
+        <Input value={storeUser.email || ''} />
+      </InlineField>
+      <InlineField label="Timezone" labelWidth={width} grow disabled>
+        <Input value={storeUser.timezone || ''} />
+      </InlineField>
+      <Legend>Notification channels</Legend>
       <Connectors {...props} />
     </>
   );